backend/requirements.txt
# Framework web
fastapi
uvicorn[standard]

# Procesamiento de datos
pandas
openpyxl
tabulate  # Necesario para pandas agent (mostrar tablas)

# LangChain con DeepSeek
langchain
langchain-experimental
langchain-community
langchain-openai  # IMPORTANTE: Necesario para DeepSeek
openai            # Cliente base de OpenAI

# Variables de entorno
python-dotenv

# Utilidades
python-multipart